Generalize conversion between persisted ontology types and records
