+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/link_type/mod.rs
@@ -11,10 +11,7 @@ use crate::{
     shared::identifier::GraphElementIdentifier,
     store::{
         crud::Read,
-        postgres::{
-            context::PostgresContext, DependencyContext, DependencyContextRef,
-            PersistedOntologyType,
-        },
+        postgres::{context::PostgresContext, DependencyContext, DependencyContextRef},
         AsClient, InsertionError, LinkTypeStore, PostgresStore, QueryError, UpdateError,
     },
     subgraph::{StructuralQuery, Subgraph},
@@ -35,7 +32,7 @@ impl<C: AsClient> PostgresStore<C> {
                 link_type_id,
                 Some(context.graph_resolve_depths.link_type_resolve_depth),
                 || async {
-                    Ok(PersistedLinkType::from_record(
+                    Ok(PersistedLinkType::from(
                         self.read_versioned_ontology_type(link_type_id).await?,
                     ))
                 },
