+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/mod.rs
@@ -6,8 +6,6 @@ mod read;
 
 use type_system::{uri::VersionedUri, DataType, EntityType, LinkType, PropertyType};
 
-pub use self::read::PersistedOntologyType;
-
 /// Provides an abstraction over elements of the Type System stored in the Database.
 ///
 /// [`PostgresDatabase`]: crate::store::PostgresDatabase
