+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/property_type/mod.rs
@@ -13,10 +13,7 @@ use crate::{
     shared::identifier::GraphElementIdentifier,
     store::{
         crud::Read,
-        postgres::{
-            context::PostgresContext, DependencyContext, DependencyContextRef,
-            PersistedOntologyType,
-        },
+        postgres::{context::PostgresContext, DependencyContext, DependencyContextRef},
         AsClient, InsertionError, PostgresStore, PropertyTypeStore, QueryError, UpdateError,
     },
     subgraph::{EdgeKind, GraphResolveDepths, OutwardEdge, StructuralQuery, Subgraph},
@@ -42,7 +39,7 @@ impl<C: AsClient> PostgresStore<C> {
                             .property_type_resolve_depth,
                     ),
                     || async {
-                        Ok(PersistedPropertyType::from_record(
+                        Ok(PersistedPropertyType::from(
                             self.read_versioned_ontology_type(property_type_id).await?,
                         ))
                     },
